Change OP_CODE to 32 bits in size, I'm an idiot. Also temporarily disable dead code warnings.

+++ src/language/operations/agents_and_teams/add_missile.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1829;
+pub const OP_CODE: u32 = 1829;
 
 pub const IDENT: &str = "add_missile";
 
 impl Operation for AddMissileOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
