+++ src/language/operations/agents_and_teams/add_reinforcements_to_entry.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1930;
+pub const OP_CODE: u32 = 1930;
 
 pub const IDENT: &str = "add_reinforcements_to_entry";
 
 impl Operation for AddReinforcementsToEntryOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
