+++ src/language/operations/agents_and_teams/agent_ai_get_behavior_target.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 2082;
+pub const OP_CODE: u32 = 2082;
 
 pub const IDENT: &str = "agent_ai_get_behavior_target";
 
 impl Operation for AgentAiGetBehaviorTargetOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
