+++ src/language/operations/agents_and_teams/agent_check_offer_from_agent.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1778;
+pub const OP_CODE: u32 = 1778;
 
 pub const IDENT: &str = "agent_check_offer_from_agent";
 
 impl Operation for AgentCheckOfferFromAgentOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
