+++ src/language/operations/agents_and_teams/agent_clear_relations_with_agents.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1802;
+pub const OP_CODE: u32 = 1802;
 
 pub const IDENT: &str = "agent_clear_relations_with_agents";
 
 impl Operation for AgentClearRelationsWithAgentsOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
