+++ src/language/operations/agents_and_teams/agent_clear_scripted_mode.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1735;
+pub const OP_CODE: u32 = 1735;
 
 pub const IDENT: &str = "agent_clear_scripted_mode";
 
 impl Operation for AgentClearScriptedModeOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
