+++ src/language/operations/agents_and_teams/agent_get_party_id.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1716;
+pub const OP_CODE: u32 = 1716;
 
 pub const IDENT: &str = "agent_get_party_id";
 
 impl Operation for AgentGetPartyIdOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
