+++ src/language/operations/agents_and_teams/agent_get_simple_behavior.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1738;
+pub const OP_CODE: u32 = 1738;
 
 pub const IDENT: &str = "agent_get_simple_behavior";
 
 impl Operation for AgentGetSimpleBehaviorOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
