+++ src/language/operations/agents_and_teams/agent_get_team.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1770;
+pub const OP_CODE: u32 = 1770;
 
 pub const IDENT: &str = "agent_get_team";
 
 impl Operation for AgentGetTeamOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
