+++ src/language/operations/agents_and_teams/agent_get_time_elapsed_since_removed.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1760;
+pub const OP_CODE: u32 = 1760;
 
 pub const IDENT: &str = "agent_get_time_elapsed_since_removed";
 
 impl Operation for AgentGetTimeElapsedSinceRemovedOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
