+++ src/language/operations/agents_and_teams/agent_get_troop_id.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1718;
+pub const OP_CODE: u32 = 1718;
 
 pub const IDENT: &str = "agent_get_troop_id";
 
 impl Operation for AgentGetTroopIdOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
