+++ src/language/operations/agents_and_teams/agent_is_defender.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1708;
+pub const OP_CODE: u32 = 1708;
 
 pub const IDENT: &str = "agent_is_defender";
 
 impl Operation for AgentIsDefenderOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
