+++ src/language/operations/agents_and_teams/agent_is_in_line_of_sight.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1826;
+pub const OP_CODE: u32 = 1826;
 
 pub const IDENT: &str = "agent_is_in_line_of_sight";
 
 impl Operation for AgentIsInLineOfSightOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
