+++ src/language/operations/agents_and_teams/agent_refill_wielded_shield_hit_points.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1692;
+pub const OP_CODE: u32 = 1692;
 
 pub const IDENT: &str = "agent_refill_wielded_shield_hit_points";
 
 impl Operation for AgentRefillWieldedShieldHitPointsOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
