+++ src/language/operations/agents_and_teams/agent_set_animation_progress.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1743;
+pub const OP_CODE: u32 = 1743;
 
 pub const IDENT: &str = "agent_set_animation_progress";
 
 impl Operation for AgentSetAnimationProgressOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
